fix: He hecho que el atributo dietas de perfil sea manytomany
He tenido que poner la clase dieta en el modelo de authentication y borrarlo del de product y hacer el import en product. Issue #118
