+++ authentication/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from authentication.models import Perfil
+from authentication.models import Perfil, Dieta
 
 # Register your models here.
 admin.site.register(Perfil)
+admin.site.register(Dieta)
