+++ authentication/models.py
@@ -1,20 +1,15 @@
 from django.db import models
 from django.contrib.auth.models import User
 
+class Dieta(models.Model):
+    nombre = models.CharField(max_length=50, null=None)
 
+    def __str__(self):
+        return self.nombre
 class Perfil(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
 
-    Dieta_Enum = (
-        ('Vegano', 'Vegano'),
-        ('Vegetariano', 'Vegetariano'),
-        ('Gluten', 'Gluten'),
-        ('Lactosa', 'Lactosa'),
-        ('Marisco', 'Marisco'),
-        ('Frutos secos', 'Frutos secos'),
-    )
-
-    dieta = models.CharField(max_length=12, choices=Dieta_Enum, default='Vegetariano', blank=False, verbose_name="Dieta")
+    dietas = models.ManyToManyField(Dieta)
 
     def __str__(self):
         return self.user.username
